requirements.txt
# Phase 0 — Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
httpx==0.28.1

# Phase 1 — RAG Pipeline
langchain==0.3.15
langchain-google-genai==2.1.4
langchain-community==0.3.15
langchain-huggingface==0.1.2
langchain-chroma==0.1.4
langchain-text-splitters==0.3.8
chromadb==0.5.3
sentence-transformers==3.3.1
pypdf==5.1.0
python-multipart==0.0.20
numpy<2

# Phase 2 — Multi-Step RAG + Evaluation
ragas==0.2.14
datasets==3.2.0

# Phase 3 — LangGraph Orchestrator
langgraph==0.2.74
tavily-python>=0.3.0

# Phase 4 — Memory
sqlalchemy>=2.0
aiosqlite

# Phase 5 — MCP Server + Email/Calendar Agent
fastmcp
mcp

# Phase 6 — Chainlit UI
chainlit
